Ajustes nos cookies da aplicação

+++ src/controllers/RestaurantsController.js
@@ -9,7 +9,12 @@ class RestaurantsController {
 
     const user = await User.findById(user_id)
     if (!user) throw new AppError('Usuário Inválido')
+    if (user.role !== 'admin') throw new AppError('Ação não permitida ao usuário!', 404)
 
+    const restaurantAlreadyExists = await Restaurant.findOne({ name: restaurantData.name, owner: user_id })
+    if (restaurantAlreadyExists) throw new AppError('Já existe um restaurante com esse nome')
+
+    restaurantData.owner = user_id
     const restaurant = await Restaurant.create(restaurantData)
 
     user.restaurants = [restaurant._id, ...user.restaurants]
@@ -21,16 +26,24 @@ class RestaurantsController {
 
   async findRestaurantsByUser(req, res) {
     const user_id = req.user.id
-
     const user = await User.findById(user_id)
+    if (!user) throw new AppError('Usuário Inválido')
 
-    const restaurantByUser = await Restaurant.find({ _id: { $in: user.restaurants } })
-
-    return res.json(restaurantByUser)
+    try {
+      const restaurantByUser = await Restaurant.find({ _id: { $in: user.restaurants } })
+      return res.json(restaurantByUser)
+    } catch (error) {
+      throw new AppError('Não foi possível buscar os restaurantes', 400)
+    }
   }
 
   async delete(req, res) {
     const restaurant_id = req.params.id
+    const user_id = req.user.id
+
+    const restaurantExists = await Restaurant.findOne({ _id: restaurant_id, owner: user_id })
+    if(!restaurantExists) throw new AppError('Restaurante não encontrado!', 404)
+
     try {
       await Restaurant.findByIdAndDelete(restaurant_id)
     } catch (error) {
